# perf: lazy load watchpack and some Node built-in modules