## Summary

- Lazy load some modules (`watchpack`, `node:vm`, `node:zlib`) to make startup faster.
- Removed an extra `new Watchpack()` to improve initial startup speed, always create watchpack instance in `the NodeWatchFileSystem.watch` method.

## Lazy load

- before:

```bash
hyperfine --warmup 10 --runs 100  'node ./dist/index.js'
Benchmark 1: node ./dist/index.js
  Time (mean ± σ):      61.4 ms ±   5.6 ms    [User: 43.2 ms, System: 16.9 ms]
  Range (min … max):    58.5 ms … 115.7 ms    100 runs
```

- after:

```bash
hyperfine --warmup 10 --runs 100  'node ./dist/index.js'
Benchmark 1: node ./dist/index.js
  Time (mean ± σ):      59.9 ms ±   0.8 ms    [User: 41.3 ms, System: 16.7 ms]
  Range (min … max):    57.8 ms …  63.2 ms    100 runs
```

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).